eld-app-backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
coreapi==2.3.3